Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/orthogonal-positioned-grid-descendants-015-expected.txt


PASS .grid 1 
PASS .grid 2 
PASS .grid 3 
PASS .grid 4 
PASS .grid 5 
PASS .grid 6 
PASS .grid 7 
PASS .grid 8 
PASS .grid 9 
FAIL .grid 10 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 3; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 3; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 11 
PASS .grid 12 
PASS .grid 13 
PASS .grid 14 
PASS .grid 15 
PASS .grid 16 
PASS .grid 17 
PASS .grid 18 
PASS .grid 19 
FAIL .grid 20 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 2; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 2; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 21 
PASS .grid 22 
PASS .grid 23 
PASS .grid 24 
PASS .grid 25 
PASS .grid 26 
PASS .grid 27 
PASS .grid 28 
PASS .grid 29 
FAIL .grid 30 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: 2; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: 2; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 31 
PASS .grid 32 
PASS .grid 33 
PASS .grid 34 
PASS .grid 35 
PASS .grid 36 
PASS .grid 37 
PASS .grid 38 
PASS .grid 39 
FAIL .grid 40 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 1; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: 1; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 41 
PASS .grid 42 
PASS .grid 43 
PASS .grid 44 
PASS .grid 45 
PASS .grid 46 
PASS .grid 47 
PASS .grid 48 
PASS .grid 49 
FAIL .grid 50 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: 1; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: 1; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 51 
PASS .grid 52 
PASS .grid 53 
PASS .grid 54 
PASS .grid 55 
PASS .grid 56 
PASS .grid 57 
PASS .grid 58 
PASS .grid 59 
FAIL .grid 60 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="145" data-offset-y="30" style="grid-column-start: 1; grid-column-end: 2; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="145" data-offset-y="30" style="grid-column-start: 1; grid-column-end: 2; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 61 
PASS .grid 62 
PASS .grid 63 
PASS .grid 64 
PASS .grid 65 
PASS .grid 66 
PASS .grid 67 
PASS .grid 68 
PASS .grid 69 
FAIL .grid 70 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: auto; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="510" data-offset-y="30" style="grid-column-start: auto; grid-column-end: auto; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 71 
PASS .grid 72 
PASS .grid 73 
PASS .grid 74 
PASS .grid 75 
PASS .grid 76 
PASS .grid 77 
PASS .grid 78 
PASS .grid 79 
FAIL .grid 80 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="445" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 3; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 81 
PASS .grid 82 
PASS .grid 83 
PASS .grid 84 
PASS .grid 85 
PASS .grid 86 
PASS .grid 87 
PASS .grid 88 
PASS .grid 89 
FAIL .grid 90 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="145" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 2; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="145" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 2; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
PASS .grid 91 
PASS .grid 92 
PASS .grid 93 
PASS .grid 94 
PASS .grid 95 
PASS .grid 96 
PASS .grid 97 
PASS .grid 98 
PASS .grid 99 
FAIL .grid 100 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="-55" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos orthogonal" data-expected-width="25" data-expected-height="0" data-offset-x="-55" data-offset-y="30" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 1; left: auto; top: 30px; right: 35px; bottom: 40px;">XX</div></div></div>
width expected 25 but got 0
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
